chore: refacor routing and store
