+++ orchestration-ui/package.json
@@ -9,7 +9,7 @@
 		"deploy": "vue-cli-service build && firebase deploy"
 	},
 	"dependencies": {
-		"axios": "^0.18.0",
+		"axios": "0.18.0",
 		"firebase": "^5.8.2",
 		"register-service-worker": "^1.5.2",
 		"vue": "^2.5.22",
