+++ orchestration-ui/src/config.js
@@ -1,0 +1,10 @@
+export default {
+	firebase: {
+		apiKey: 'AIzaSyCvBMLoUZafodd2vlyAxUs3Pl0mEARUOEc',
+		authDomain: 'holistic-home-5134d.firebaseapp.com',
+		databaseURL: 'https://holistic-home-5134d.firebaseio.com',
+		projectId: 'holistic-home-5134d',
+		storageBucket: 'holistic-home-5134d.appspot.com',
+		messagingSenderId: '549410338560'
+	}
+};
