+++ orchestration-ui/src/main.js
@@ -4,7 +4,7 @@ import Vuetify from 'vuetify';
 import 'vuetify/dist/vuetify.min.css';
 import App from './App.vue';
 import router from './router';
-import storeConfig from './vuex/store';
+import store from './store/index';
 import theme from './styles/theme';
 
 
@@ -12,8 +12,6 @@ Vue.config.productionTip = false;
 
 Vue.use(Vuex);
 Vue.use(Vuetify, { theme });
-
-const store = new Vuex.Store(storeConfig);
 
 new Vue({
 	router,
