+++ orchestration-ui/src/router/routes.js
@@ -1,13 +1,12 @@
 import accountRoutes from '../containers/Account';
 import controlRoutes from '../containers/Control';
 
-
 const routes = [
 	...accountRoutes,
 	...controlRoutes,
 	{
 		path: '*',
-		redirect: { name: 'account.unauthorized' }
+		redirect: { name: 'control.Page' }
 	}
 ];
