--- orchestration-ui/src/store.js
@@ -1,16 +1,0 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
-
-Vue.use(Vuex);
-
-export default new Vuex.Store({
-	state: {
-
-	},
-	mutations: {
-
-	},
-	actions: {
-
-	}
-});
