+++ orchestration-ui/src/store/lib/buildCollectionStore.js
